move problem page to feature, rename getProblemInfo -> getProblem
