+++ .eslintrc.cjs
@@ -16,7 +16,7 @@ module.exports = {
     "no-restricted-imports": [
       "error",
       {
-        patterns: ["@/providers/*/*", "@/hooks/*", "@/redux/*/*", "@/types/*"],
+        patterns: ["@/providers/*/*", "@/hooks/*", "@/redux/*/*", "@/types/*", "@/features/*/*"],
       },
     ],
     "@typescript-eslint/ban-ts-comment": "off",
