+++ src/app/problems/[pid]/page.tsx
@@ -1,20 +1,16 @@
-import { getProblemInfo } from "@utils/api";
+import { getProblem } from "@utils/api";
 import type { Metadata } from "next";
 import { Suspense } from "react";
 
 import { Box, Heading } from "@/components";
-
-import { Editor } from "./Editor";
-import { InfoTable } from "./InfoTable";
-import { InfoTableScore } from "./InfoTableScore";
-import { RecalcScore } from "./RecalcScore";
+import { Editor, InfoTable, RecalcScore } from "@/features/problems";
 
 export const metadata: Metadata = {
   title: "Problem",
 };
 
 const Page = async ({ params: { pid } }: { params: { pid: string } }) => {
-  const problemData = await getProblemInfo(pid);
+  const problemData = await getProblem(pid);
 
   if (!problemData.ok) {
     console.error(problemData.status, problemData.error);
@@ -31,21 +27,7 @@ const Page = async ({ params: { pid } }: { params: { pid: string } }) => {
   return (
     <div className="mx-auto flex w-full flex-col items-start gap-6 self-stretch px-3 py-5 md:max-w-7xl md:flex-row md:gap-4 md:px-2 md:py-10 lg:gap-8 lg:px-4">
       <div className="h-auto w-full md:flex-[1_1_0]">
-        <InfoTable
-          score={
-            <Suspense
-              fallback={
-                <div className="pb-4 pt-5">
-                  <div className="mx-auto h-8 w-28 animate-pulse rounded-md bg-gray-200 dark:bg-slate-600"></div>
-                </div>
-              }
-            >
-              <InfoTableScore problemId={pid} />
-            </Suspense>
-          }
-          problemData={problemData.data}
-          pid={pid}
-        />
+        <InfoTable problemData={problemData.data} pid={pid} />
         <RecalcScore dpid={problemData.data.dpid} />
       </div>
       <div className="flex h-full w-full md:flex-[2_2_0]">
