+++ src/features/problems/components/ProblemPage/Editor/index.tsx
@@ -1,6 +1,6 @@
 import { cookies } from "next/headers";
 
-import { EditorClient, type EditorClientProps } from "./EditorClient";
+import { EditorClient, type EditorClientProps } from "./Client";
 
 export const Editor = async (props: Omit<EditorClientProps, "isLoggedIn">) => {
   const cookiesStore = cookies();
