+++ src/features/problems/components/ProblemPage/index.ts
@@ -1,0 +1,3 @@
+export * from "./Editor";
+export * from "./InfoTable";
+export * from "./RecalcScore";
