+++ src/features/problems/queries/index.ts
@@ -1,0 +1,1 @@
+export * from "./getProblem";
