+++ src/shared/utils/api/index.ts
@@ -1,7 +1,7 @@
 import "server-only";
+export { getProblem } from "../../../features/problems/queries/getProblem";
 export { getAllProblem } from "./getAllProblem";
 export { getMemoProblemsScore } from "./getMemoProblemsScore";
 export { getProblemCount } from "./getProblemCount";
-export { getProblemInfo } from "./getProblemInfo";
 export { type DetailedScoreInfo, getProblemScore } from "./getProblemScore";
 export { getSubmissionInfo, type SubmissionInfo } from "./getSubmissionInfo";
